+++ src/inspect_ai/log/_transcript.py
@@ -111,7 +111,7 @@ class SampleLimitEvent(BaseEvent):
     message: str
     """A message associated with this limit"""
 
-    limit: float | None = Field(default=None)
+    limit: int | None = Field(default=None)
     """The limit value (if any)"""
